+++ apmconfig/doc.go
@@ -15,19 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package transport
-
-import (
-	"context"
-	"io"
-)
-
-// Transport provides an interface for sending streams of encoded model
-// entities to the Elastic APM server, and for querying config. Methods
-// are not required to be safe for concurrent use.
-type Transport interface {
-	// SendStream sends a data stream to the server, returning when the
-	// stream has been closed (Read returns io.EOF) or the HTTP request
-	// terminates.
-	SendStream(context.Context, io.Reader) error
-}
+// Package apmconfig provides an API for watching agent config
+// changes.
+package apmconfig