+++ apmtest/configwatcher.go
@@ -15,19 +15,18 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package transport
+package apmtest
 
 import (
 	"context"
-	"io"
+
+	"go.elastic.co/apm/apmconfig"
 )
 
-// Transport provides an interface for sending streams of encoded model
-// entities to the Elastic APM server, and for querying config. Methods
-// are not required to be safe for concurrent use.
-type Transport interface {
-	// SendStream sends a data stream to the server, returning when the
-	// stream has been closed (Read returns io.EOF) or the HTTP request
-	// terminates.
-	SendStream(context.Context, io.Reader) error
+// WatchConfigFunc is a function type that implements apmconfig.Watcher.
+type WatchConfigFunc func(context.Context, apmconfig.WatchParams) <-chan apmconfig.Change
+
+// WatchConfig returns f(ctx, params).
+func (f WatchConfigFunc) WatchConfig(ctx context.Context, params apmconfig.WatchParams) <-chan apmconfig.Change {
+	return f(ctx, params)
 }